spilt test file
