+++ packages/zenscript/test/zenscript/top-level.test.ts
@@ -1,43 +1,14 @@
-import { parseHelper } from 'langium/test'
-import { NodeFileSystem } from 'langium/node'
-import { describe, expect, it, suite } from 'vitest'
+import { describe, expect, it } from 'vitest'
 
-import type { LangiumDocument } from 'langium'
-import type { PrimitiveType, ReferenceType, Script, TypeReference } from '../src/generated/ast'
-import { createIntelliZenServices } from '../src/module'
+import { assertNoErrors, createParseHelper, expectTypeToBe } from '../utils'
 
-const service = createIntelliZenServices(NodeFileSystem)
-const parse = parseHelper<Script>(service.intelliZen)
+const parse = createParseHelper()
 
 async function parseModel(input: string) {
   return parse(input, { validation: true })
 }
 
-async function assertNoErrors(model: LangiumDocument<Script>) {
-  expect(model.parseResult.lexerErrors).toHaveLength(0)
-  expect(model.parseResult.parserErrors).toHaveLength(0)
-
-  if (model.diagnostics?.[0]?.data?.code !== 'linking-error') {
-    expect(model.diagnostics ?? []).toHaveLength(0)
-  }
-}
-
-function expectTypeToBe(matches: string, type?: TypeReference) {
-  if (type?.$type === 'PrimitiveType') {
-    expect((type as PrimitiveType).value).toBe(matches)
-  }
-  else if (type?.$type === 'ReferenceType') {
-    expect((type as ReferenceType).ref.$refText).toBe(matches)
-  }
-}
-
-describe('test with langium syntax parse', async () => {
-  it('missing semicolon', async () => {
-    const model = await parseModel('val a as int')
-    expect(model.parseResult.lexerErrors).toHaveLength(0)
-    expect(model.parseResult.parserErrors).toHaveLength(1)
-  })
-
+describe('parse ZenScript top-level of script ', () => {
   it('import declaration', async () => {
     const model = await parseModel('import foo.bar.baz;')
     const refImport = model.parseResult.value.imports[0]
